+++ docs/source/release_notes.rst
@@ -13,6 +13,9 @@ Release Notes
 dev
 ---
 
+**New features**
+ - Add possibility in schema to specify whether a variable length container (list or dict) is allowed to be empty
+
 0.6.4 (2020-09-25)
 ------------------